updated with feedback from review
